requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
altair==4.2.2
    # via streamlit
attrs==23.1.0
    # via jsonschema
blinker==1.6.2
    # via streamlit
cachetools==5.3.0
    # via streamlit
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
decorator==5.1.1
    # via validators
deta==1.1.0
    # via -r requirements.in
entrypoints==0.4
    # via altair
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
idna==3.4
    # via requests
importlib-metadata==6.6.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.24.3
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==23.1
    # via streamlit
pandas==1.5.3
    # via
    #   altair
    #   streamlit
pillow==9.5.0
    # via streamlit
protobuf==3.20.3
    # via streamlit
pyarrow==11.0.0
    # via streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.28.2
    # via streamlit
rich==13.3.4
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
streamlit==1.21.0
    # via -r requirements.in
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.1
    # via streamlit
tqdm==4.65.0
    # via -r requirements.in
typing-extensions==4.5.0
    # via streamlit
tzdata==2023.3
    # via
    #   pytz-deprecation-shim
    #   tzlocal
tzlocal==4.3
    # via streamlit
urllib3==1.26.15
    # via requests
validators==0.20.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
zipp==3.15.0
    # via importlib-metadata
